Move template.go to util package
- This commit moves the template.go from pkg/components/utl to pkg/util

- Update package references for pkg/components to include new package
  location
