+++ pkg/components/prometheus-operator/component.go
@@ -23,6 +23,7 @@ import (
 
 	"github.com/kinvolk/lokomotive/pkg/components"
 	"github.com/kinvolk/lokomotive/pkg/components/util"
+	utilpkg "github.com/kinvolk/lokomotive/pkg/util"
 )
 
 const name = "prometheus-operator"
@@ -95,7 +96,7 @@ func (c *component) RenderManifests() (map[string]string, error) {
 		return nil, errors.Wrap(err, "load chart from assets")
 	}
 
-	values, err := util.RenderTemplate(chartValuesTmpl, c)
+	values, err := utilpkg.RenderTemplate(chartValuesTmpl, c)
 	if err != nil {
 		return nil, errors.Wrap(err, "render chart values template")
 	}
